requirements.txt
# ============================================================
#  requirements.txt
# ============================================================

# OpenAI SDK
openai>=1.30.0

# .env 파일 로드 (API 키 보안 관리)
python-dotenv>=1.0.0

# 패킷 스니핑 (Kali 전용 — Windows에서는 불필요)
scapy>=2.5.0

# Gradio 대시보드
gradio>=4.36.0
